+++ lib/sidekiq/throttled/registry.rb
@@ -102,8 +102,6 @@ module Sidekiq
         # @param name [Class, #to_s]
         # @return [Strategy, nil]
         def find_by_class(name)
-          return unless Throttled.configuration.inherit_strategies?
-
           const = name.is_a?(Class) ? name : Object.const_get(name)
           return unless const.is_a?(Class)
 
@@ -112,6 +110,8 @@ module Sidekiq
             return strategy if strategy
           end
 
+          nil
+        rescue NameError
           nil
         end
       end