+++ CHANGES.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Correctly finalize throttled jobs when used with ActiveJob
+  [#151](https://github.com/ixti/sidekiq-throttled/pull/151)
+
 
 ## [1.0.0.alpha.1] - 2023-06-08