+++ CHANGES.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Correctly finalize throttled jobs when used with ActiveJob
   [#151](https://github.com/ixti/sidekiq-throttled/pull/151)
 
+### Removed
+
+- (BREAKING) Drop Ruby-2.7.x support
+
 
 ## [1.0.0.alpha.1] - 2023-06-08